add new function to Weave and add change nrow to usize

+++ src/data.rs
@@ -77,11 +77,11 @@ fn cast_field_to_number<D: Number>(field: &Field) -> D {
     }
 }
 
-pub fn read_parquet_nrow<P>(path: P) -> GenResult<i32>
+pub fn read_parquet_nrow<P>(path: P) -> GenResult<usize>
 where
     P: AsRef<Path>,
 {
     let file = File::open(path)?;
     let reader = SerializedFileReader::new(file)?;
-    Ok(reader.metadata().file_metadata().num_rows() as i32)
+    Ok(reader.metadata().file_metadata().num_rows() as usize)
 }
